# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Metron Screen-Time Orchestration API
  description: |
    Metron is a unified screen-time orchestration system that tracks children's screen time
    across multiple devices (TV, PS5, iPad, etc.) with driver-based integrations.

    This API provides endpoints for managing children, sessions, devices, and viewing statistics.

    ## Features
    - Multi-child session management
    - Device driver integration (Aqara Cloud, PS5, etc.)
    - Break rule enforcement
    - Daily usage tracking
    - Real-time session control

    ## Authentication
    All `/v1/*` endpoints require API key authentication via the `X-Metron-Key` header.
    The `/health` endpoint does not require authentication.

  version: 1.0.0
  contact:
    name: Metron API Support
    url: https://github.com/4rakcheev/metron
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://metron.example.com
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/devices/{id}/bypass:
    post:
      tags:
        - Bypass
      summary: Set device bypass mode
      description: |
        Enable or disable bypass mode for a device. When bypass is active,
        agents will skip enforcement and allow unrestricted access.

        Bypass can be set with an optional expiration time.
      operationId: setDeviceBypass
      parameters:
        - name: id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            example: win-pc1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetBypassRequest'
            examples:
              enableWithExpiry:
                summary: Enable bypass for 1 hour
                value:
                  enabled: true
                  reason: homework
                  expires_in_minutes: 60
              enableIndefinite:
                summary: Enable indefinite bypass
                value:
                  enabled: true
                  reason: special occasion
              disable:
                summary: Disable bypass
                value:
                  enabled: false
      responses:
        '200':
          description: Bypass mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BypassStatus'
              examples:
                enabled:
                  value:
                    device_id: win-pc1
                    enabled: true
                    reason: homework
                    enabled_at: "2025-12-09T15:30:45Z"
                    expires_at: "2025-12-09T16:30:45Z"
                disabled:
                  value:
                    device_id: win-pc1
                    enabled: false
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Bypass
      summary: Clear device bypass mode
      description: Remove bypass mode for a device, re-enabling normal enforcement.
      operationId: clearDeviceBypass
      parameters:
        - name: id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            example: win-pc1
      responses:
        '204':
          description: Bypass cleared successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Metron-Key
      description: API key for admin authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for agent authentication. Each token is tied to a specific device.
  schemas:
...
    Error:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Session not found
        code:
          type: string
          description: Machine-readable error code
          example: SESSION_NOT_FOUND
        details:
          type: string
          description: Additional error details (optional)
          example: Invalid JSON in request body
...
    SetBypassRequest:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether to enable or disable bypass
          example: true
        reason:
          type: string
          description: Optional reason for enabling bypass
          example: homework
        expires_in_minutes:
          type: integer
          description: Minutes until bypass expires. Omit for indefinite bypass.
          minimum: 1
          example: 60
    BypassStatus:
      type: object
      required:
        - device_id
        - enabled
      properties:
        device_id:
          type: string
          description: Device ID
          example: win-pc1
        enabled:
          type: boolean
          description: Whether bypass is enabled
          example: true
        reason:
          type: string
          description: Reason for bypass (only present if enabled)
          example: homework
        enabled_at:
          type: string
          format: date-time
          description: When bypass was enabled (only present if enabled)
          example: "2025-12-09T15:30:45Z"
        expires_at:
          type: string
          format: date-time
          description: When bypass expires (only present if enabled with expiry)
          example: "2025-12-09T16:30:45Z"
  responses:
    UnauthorizedError:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: UNAUTHORIZED
    BadRequestError:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidRequest:
              summary: Invalid request body
              value:
                error: Invalid request body
                code: INVALID_REQUEST
                details: "minutes must be positive"
            insufficientTime:
              summary: Insufficient time remaining
              value:
                error: Child has insufficient remaining time
                code: INSUFFICIENT_TIME
            invalidAction:
              summary: Invalid action
              value:
                error: Invalid action. Must be 'extend' or 'stop'
                code: INVALID_ACTION
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            code: INTERNAL_ERROR